requirements.txt
# requirements.txt
camelot-py[cv]
pdfplumber
pandas
pytesseract
opencv-python
fastapi
uvicorn
pydantic
python-multipart

